refactor(state): relocate classifyHealth into the State surface, unchanged (#266)
